Unify renderer, renderer config and templates under app module
