+++ src/lib/app/renderer/template.rs
@@ -1,36 +1,6 @@
-use crate::app::config::render::RendererConfig;
-use actix::{Actor, Addr, Handler, Message, SyncArbiter, SyncContext};
+use super::RendererInner;
+use actix::{Handler, Message};
 use serde_json::Value;
-use std::ops::Deref;
-use std::path::Path;
-use std::sync::{Arc, Mutex};
-use tera::Tera;
-
-#[derive(Clone)]
-pub struct Renderer(pub Addr<RendererInner>);
-
-impl Renderer {
-    pub fn new(conf: RendererConfig) -> Self {
-        let path = Path::new(&conf.templates).join("**/*");
-        let tera = Arc::new(Mutex::new(Tera::new(path.to_str().unwrap()).unwrap()));
-
-        Self(SyncArbiter::start(3, move || RendererInner(tera.clone())))
-    }
-}
-
-impl Deref for Renderer {
-    type Target = Addr<RendererInner>;
-
-    fn deref(&self) -> &Addr<RendererInner> {
-        &self.0
-    }
-}
-
-pub struct RendererInner(pub Arc<Mutex<Tera>>);
-
-impl Actor for RendererInner {
-    type Context = SyncContext<Self>;
-}
 
 pub struct Template {
     pub name: String,
