+++ src/lib/util/mod.rs
@@ -1,4 +1,3 @@
 pub mod command;
 pub mod config;
-pub mod render;
 pub mod shell;
